@@ -50,16 +50,16 @@
     "workspace-generator": "nx workspace-generator"
   },
   "dependencies": {
-    "@angular/animations": "13.3.6",
-    "@angular/cdk": "13.3.6",
-    "@angular/common": "13.3.6",
-    "@angular/compiler": "13.3.6",
-    "@angular/core": "13.3.6",
-    "@angular/forms": "13.3.6",
-    "@angular/material": "13.3.6",
-    "@angular/platform-browser": "13.3.6",
-    "@angular/platform-browser-dynamic": "13.3.6",
-    "@angular/router": "13.3.6",
+    "@angular/animations": "14.0.2",
+    "@angular/cdk": "14.0.1",
+    "@angular/common": "14.0.2",
+    "@angular/compiler": "14.0.2",
+    "@angular/core": "14.0.2",
+    "@angular/forms": "14.0.2",
+    "@angular/material": "14.0.1",
+    "@angular/platform-browser": "14.0.2",
+    "@angular/platform-browser-dynamic": "14.0.2",
+    "@angular/router": "14.0.2",
     "@codewithdan/observable-store": "2.2.11",
     "@dinero.js/currencies": "2.0.0-alpha.8",
     "@nestjs/bull": "0.5.5",
@@ -166,15 +166,15 @@
     "import-sort-parser-typescript": "6.0.0",
     "import-sort-style-module": "6.0.0",
     "jest": "27.5.1",
-    "jest-preset-angular": "11.1.1",
-    "nx": "14.1.4",
-    "prettier": "2.5.1",
+    "jest-preset-angular": "11.1.2",
+    "nx": "14.3.5",
+    "prettier": "2.7.1",
     "replace-in-file": "6.2.0",
     "rimraf": "3.0.2",
     "tslib": "2.0.0",
     "ts-jest": "27.1.4",
-    "ts-node": "9.1.1",
-    "typescript": "4.6.4"
+    "ts-node": "10.8.1",
+    "typescript": "4.7.3"
   },
   "engines": {
     "node": ">=14"
