@@ -51,10 +51,10 @@ export class FireCalculatorComponent
   @ViewChild('chartCanvas') chartCanvas;
 
   public calculatorForm = this.formBuilder.group({
-    annualInterestRate: new FormControl(),
-    paymentPerPeriod: new FormControl(),
-    principalInvestmentAmount: new FormControl(),
-    time: new FormControl()
+    annualInterestRate: new FormControl<number>(undefined),
+    paymentPerPeriod: new FormControl<number>(undefined),
+    principalInvestmentAmount: new FormControl<number>(undefined),
+    time: new FormControl<number>(undefined)
   });
   public chart: Chart;
   public isLoading = true;
@@ -261,15 +261,13 @@ export class FireCalculatorComponent
       this.calculatorForm.get('principalInvestmentAmount').value || 0;
 
     // Payment per period
-    const PMT: number = parseFloat(
-      this.calculatorForm.get('paymentPerPeriod').value
-    );
+    const PMT = this.calculatorForm.get('paymentPerPeriod').value;
 
     // Annual interest rate
     const r: number = this.calculatorForm.get('annualInterestRate').value / 100;
 
     // Time
-    const t: number = parseFloat(this.calculatorForm.get('time').value);
+    const t = this.calculatorForm.get('time').value;
 
     for (let year = currentYear; year < currentYear + t; year++) {
       labels.push(year);
